Added room_first, room_last and fixed _pattern
-Added room_first and room_last. They were declared, but not
implemented. So they trew undefined reference when used.
-Fixed draw_set_line_pattern having wrong definition, so it also trew
undefined reference erro when used.

+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DX11draw.cpp
@@ -78,7 +78,7 @@ void draw_set_alpha_test_ref_value(unsigned val)
 
 }
 
-void draw_set_line_pattern(unsigned short pattern, int scale)
+void draw_set_line_pattern(int pattern, int scale)
 {
 
 }
